backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.38.0
pydantic==2.12.5
google-generativeai==0.8.3
aiohttp==3.11.11
python-multipart==0.0.20
Pillow==11.1.0
pypdf==5.1.0
aiosqlite==0.20.0
psutil==6.1.1
cryptography==44.0.0

# QR Code Generation
qrcode[pil]>=7.4.2

# Optional dependencies for revolutionary features
# Install with: pip install chromadb sentence-transformers gitpython sqlalchemy psycopg2-binary pymysql playwright

# RAG Document Intelligence
chromadb>=0.4.22
sentence-transformers>=2.2.0
python-docx>=1.0.0
markdown>=3.5.0

# Git Integration
# gitpython>=3.1.0

# Database Query Builder
# sqlalchemy>=2.0.0
# psycopg2-binary>=2.9.0
# pymysql>=1.1.0

# Browser Automation
# playwright>=1.40.0

# Google Integration (already optional)
# google-auth>=2.0.0
# google-auth-oauthlib>=1.0.0
# google-api-python-client>=2.0.0
# PyAudio requires system dependencies. Install with:
# Ubuntu/Debian: sudo apt-get install portaudio19-dev python3-pyaudio
# Then: pip install pyaudio
numpy>=1.26.0
speechrecognition==3.10.0
pyttsx3==2.90
python-dotenv==1.0.0
google-generativeai
Pillow>=10.0.0
pypdf>=3.17.0
aiosqlite>=0.19.0
